xrequest update - react version
Now properly supports react's props management.
Use downloadSource() in the main App.js to get the route definitions file, and then save it in state. Pass the contents as props to other components if they need to perform API calls.
User setSource() to get the route definitions out of an object variable from props in any component. Set this in the OnComponentMount() method.